Added alert and error handling

+++ src/app/user/watchlist/watchlist.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from '../models/product.model';
 import { WatchlistService } from '../services/watchlist/watchlist.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-watchlist',
@@ -22,9 +23,24 @@ export class WatchlistComponent implements OnInit {
   }
 
   removeProduct(id: number): void {
-    this.watchlistService.removeProductFromWatchlist(id).subscribe(() => {
-      this.products = this.products.filter(product => product.productId !== id);
-    })
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You won’t be able to revert this',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, remove the product!',
+      cancelButtonText: 'Cancel',
+      confirmButtonColor: '#3085d6', 
+      cancelButtonColor: '#d33'    
+    }).then(result => {
+      if (result.isConfirmed) {
+        this.watchlistService.removeProductFromWatchlist(id).subscribe(() => {
+          this.products = this.products.filter(product => product.productId !== id);
+        })
+      }
+    });
+
+
   }
 
 }
